Drop predicates
Add support for custom drop predicates. Each type has a predicate
generated which captures the assumptions generated from dropping a value
of this type. When a variable is dropped we assume the predicate holds,
gaining the relevant assumptions.

This also involved some refactoring of parts of the code, and fixing a
few bugs in the printer.