Detect and forbid mutually recursive types
This commit adds a check while translating type defintions to ensure
that we are not translating mutually recursive types. In the future this
restriction could be lifted without too much difficulty as 'most' of the
hard parts are done.

It also introduces a context type for type translation. This will help
if future passes and work need to introduce additional context.