Initial version of specification language
This rather large commit contains an initial sketch of the specification
language implementation for creusot. It reorganizes the layout of the
contract related crates, properly masking the proc macros behind a
normal crate facade. It also adds the `pearlite` crate which contains
the implementation of the specification language for creusot.

Currently, pearlite has a parser from my custom fork of syn into its own
ast as well as a basic typechecker. In the future I'd like to merge the
syn-parser with the pearlite parser but it isn't currently possible as I
need several private syn apis to properly parse rust-style code.

Creusot internally translates pearlite to Why3 code and provides an
interface for the pearlite typechecker to access Rust functions and type
declarations