Cleanup name translation somewhat
More cleanly distinguish when we want to translate a defid as a value
versus when we want to translate it as a type. This is important to
handle structs in Rust as they can be both their type and constructor
simultaneously.

Additionally, to make things simpler when we lower a
pearlite path to mlcfg we throw away all the strings and rebuild the
name from scratch rather than attempt to map the strings over. This
might not be efficient but who cares, it's clearly not the only issue in
the code.