Extract common context type
Gets rid of `ty::Ctx` in favor of a `TranslationCtx` which tracks all
the global information related to translation such as the already
translated ids and the output declarations.

This change also affects the printing of modules, cleaning it up a bit
but requiring the test cases to be updated as a result.