Drop old values during assignments
This commit fixes the translation of assignments and adds support for
`replace`, which is essentially a drop+assignment.

The problem with the current translation of assignments was that it was
too weak. When a value is assigned, we don't resolve the *previous*
value that was contained.

This commit fixes this by resolving the value during `replace` and
during a `Use` assignment.
It feels like it could be important to do it during a `Ref` assignment,
but i would like to check first.

Because this changes the translation of every program which mutates a
variable, the tests are updated.