Remove poorly thought-out `SimplePlace` type
The SimplePlace type was a poor wrapper around rustc's Place type which
provided alternate projections that were 'closer' to what Creusot
needed. However, in practice, we ended up needing more / different
information than what those projections provided (see handling of
`FieldAccess`). It also was blocking the implementation of certain
features where we need access to the type *pointed* to by a `Place`,
something that was not possible with `SimplePlace`.

For now I've chosen to operate directly on the `Place` type, and with
time perhaps a better barrier / representation will become apparent.