Structure output as flat list of modules
Why3 doesn't support nesting modules and namespaces, and since functions
are now translated to modules, it makes more sense to print them as a
flat list. The name of each module will be the full concatenated path
from Rust.

This also means that we must import the prelude / Type module each time.

We also address a long-standing bug/issue by generating opaque values
for external functions that are used. They have no contract but allow
programs which call external functions (ie abort) to at the very least
be syntactically valid.

It's not 100% clear what the proper solution is for multi-crate
projects, but that will be dealt with when we get there.