Rewrite Pearlite typechecker
I've chosen to throw away the older pearlite typechecker which was
proving challenging to extend to a more complete language. In particular
I wanted to add support for traits and polymorphic functions in Pearlite
but this would have required reimplementing most of the Rust type
system. While certain crates like Chalk help in making this simpler,
tasks like name resolution remain incredibly challenging.

My solution is to instead translate pearlite terms in a HOAS style, and
type check them using the *actual* Rust typechecker. This allows
Pearlite to get traits for 'free' from Rust and has the added benefit of
nice error messages from Rust.

However, since Pearlite terms are 'logical' we don't want to restrict
them to Rust's ownership discipline, so we intercept the compilation
before the borrowchecker is run, meaning that pearlite times do not have
move / copy / borrow restrictions at all!

In the long term this may be limiting if Pearlite grows features which
don't exist or are incompatible with Rust's type system. Additionally, I
feel like error messages could be improved by being specifically adapted
to pearlite, especially since it is actually a *subset* of Rust but the
compiler doesn't know that.