Initial support for traits
This commit adds the initial version of support for traits to Creusot.
In this first version, we can translate traits, their instances and use
sites. Traits can also contain logic functions, and so we've overhauled
the half-broken logical resolution mechanism that was present. The
previous version would generate ill-formed code when using generics.

Each trait is translated to a *module* in Why3, and instances clone this
module, substituting the correct implementations for various functions.
Each use site then either clones the trait or the concrete
implementation which is being used.

This support is initial, there is still a lot of work to do to fill in
gaps and crashes. In particular:

* Support for supertraits is rough, the translation is *correct* but
  the substitution is too weak. Currently it doesn't include supertrait
  methods, meaning that Why3 won't understand the links between a
  supertrait and its descendants.
* Support for associated types & constants. These were excluded, but
  there is no forseen difficulty in adding them in.
* "Specialization" of trait methods. When possible we should refine
  calls to the *concrete* method being called while stil bringing into
  scope the overall instance / trait. This allows individual instances
  to *strengthen* trait contracts.
* Marker trait inference. Marker traits in Rust have no methods of their
  own (think `Eq` vs `ParialEq`) but they guarantee additional
  properties about their supertraits. Since they never appear in the
  code (no method calls), we may not pick up that there is an instance
  for that marker trait available which may be necessary for a proof.
  For example, to use reflexivity of `==` we would need to know we can
  bring in an `Eq` instance. At the moment you must explicitly specify
  all the instances you want in scope, in the future we may add some
  amount of inference.