Support for multi-crate projects and dependencies
With the new typechecker, using logical functions and predicates
requires storing some form of metadata which can be read across crates.
In particular the upcoming work on traits will require this to perform
logical resolution.

This commit lays out all the groundwork for this and along the way
changes the binary interface of creusot. Gone is `main.rs`, replaced
with three binaries that work togehter: `cargo-creusot` a cargo
extension, `creusot-rustc` a rustc wrapper, and `creusot-driver` the
actual rustc driver which does the work.

When creusot runs, it will emit metadata about all the exported symbols,
in particular their translated representations. This allows dependent
crates to load this metadata and recover the pre-translated function.
This is only done for logical and predicate functions for the moment,
though it will also be done for all public function signatures in the
future.

As part of these changes creusot also gained a few new options to toggle
various parts of the metadata loading and storing which are mainly
useful for tests.