Initial (broken) version of associated types
This commit includes an initial, but broken version of associated types
in traits. In particular they do not interact well with super traits /
trait bounds :/. However, we even this broken version can function in
simple cases and will simply refuse to load in why3 otherwise (ie: not
silently fuck up proofs). While I continue fixing the underlying issues
I'm merging this to unblock the `creusot-contracts` library which was
failing to build otherwise.