Overhaul management of dependencies
Currently, there is an issue where we will duplicate function
definitions if they are used multiple times in the transitive call
graph. Solving this required a complete overhaul of the function
management of Creusot.

The idea in short is to give each function logical and program an
interface as well as its implementation. Then it becomes a matter of
`clone`ing the correct version of each function. This last point is
rather subtle because we want logic functions to be *transparent* at
their use-sites, that is we want access to the full body of the logic
function. So here is how it works:

- A logic function will clone the *interface* of any logic function it
  uses. This means that the definition of a logic function is *abstract*
  over the bodies of definitions.
- A program function will clone the *interface* of other *program*
  functions, on the other hand it clones the *implementation* of logic
  functions. This allows logic functions to be transparent when
  performing the proofs of program functions.