Rework the invariant mover
Because invariants are embedded as closures in the MIR code, just before
a loop starts, we need to push them down into the actual loop header
block. The previous code always assumed that this was the following
block, but in general there can be a series of 'dummy blocks' before we
hit the right one.

Additionally, we must 'gather' invariants *before* getting the MIR for
the function we are translating. This is because we translate invariants
from *THIR* instead, and we must get that THIR before it gets stolen
during MIR generation.

This commit reworks and unifies these two operations, while also
providing much more robust invariant moving logic. We push the
invariants down until we encounter a loop header, which we determine as
a block which dominates a predecessor (standard definition).