Improve cloning of pure functions
Ensure that the interface of a pure function only contains the
`function` symbol of the function rather than its value which has a
contract. The interface of a pure function is only used in other
interfaces or the bodies of logical functions, thus, we never need the
program symbol in it.

The program symbol can only be used from other program functions which
don't use the interface anyways.