Add impl interfaces.
Like logic functions the module of an impl can easily
appear in both a function interface and body. This occurs when we
instantiate a constraint of a dependency with a concrete implementation.
As a result, we must ensure that impls have an interface as well as a
body module.

However, even the interface of an impl will still fully define
associated types, which in turn means we must ensure that we don't
attempt to further refine them.