Fix implementation refinement
Restores the ability for implementations to refine the contracts of
traits. This required a complete overhaul of how we emit pre/post
conditions: they must be emitted as closures instead of top level
functions.

Since we no longer emit `fn`s we cannot use `rustc_diagnostic_item` to
find the interesting definitiosn. Thus, I introduced `creusot::item`
which works in the same manner but can be attached to anything.
This required extending the crate metadata, so I took advantage of the
moment to do some cleanups there.