Overhaul trait and impl handling
Rather than emitting a single module for traits and implementations, we
now emit one module per item belonging to the trait or impl. This means
that when we clone a trait/impl member we directly clone the module
which provides it.

This also solves #30, as we can now easily clone multiple instances of a
trait member with different parameters without causing the whole trait
to get brought in.