Make ghost a normal type
In Creusot we really want to use `Ghost` as a normal type, even if that
means occasionally confusing things like `Ghost<Ghost<T>>`, in exchange
we can do things like have ghost parameters to closures. This PR removes
all usage of Why3-level ghost code and instead replaces it with a
`Ghost` type which serves purely as a syntactic marker / wrapper for
ghost code.