Split deps out from TranslatedItems
This splits dependencies away from translated items, so we can
store the dependencies for a function before recursively translating
function it calls.

See discussion in #552, but while this does fix an issue where
dependencies aren't properly substituted in a cloned theorem, it
doesn't fix the the main ordering issue in #552.