Better encoding of spans: use already imported files instead of loading new files, and support SyntaxtContext serialization.
A lot of the encoding/decoding machinery is taken from rustc.

Also, removed a test that made imported never used for spans in generated mlcfg files.
For span comming from the stdlib, if a relative path is required, we don't show any span info, because the stdlib is unlikely to be installed in a relatively stable directory.