fix: emit correct relative paths in spans with 'cargo creusot'
add an option to creusot-rustc to specify the relative location of the
project's root with respect to creusot's output. This option is used
by the test harness, which uses a different file layout than 'cargo
creusot'. ('cargo creusot' then simply uses the default value of the
option, which matches its usecase.)