New VC for lemma functions
This commit completely reworks the verification of lemma (logic)
functions. In particular, they no longer generate why3's `let functions`
which turns out are very problematic when combined with other features
of why3 like cloning and totally incompatible with `Coma`.

The approach is based off of WP-calulus though several adaptations have
to be made given that the terms being verified are logical formulas not
programs.