Rework/fix the computation of relative spans
The handling of file paths and the logic for computing relative spans
was subtly broken/inconsistent in a few places.

- --root-path-relative-from-output really only made sense when passing
  --stdout. Otherwise, we know the path of the output mlcfg file, we
  don't need it to compute relative spans, and we don't know what
  to do with it. Furthermore the concept of the "root" of the
  project seems dubious outside of cargo (but this option is used
  in creusot-rustc).

- the logic for computing relative spans in
  creusot::options::relative_to_output was therefore somewhat
  nonsensical.

This replaces the option with --spans-relative-to which instead
indicates a directory path with respect to which spans should be
relative to. We then require this option to be passed whenever
--stdout is passed and we're using relative spans.

Then, creusot::options::relative_to_output can be simply replaced by a
standard "pathdiff" implementation from a crate.

As a extra simplification, in creusot-rustc we require to receive one
of --stdout or --output-file, which means we can remove any
path-computation logic for .mlcfg from creusot-rustc (everything is
done by cargo-creusot).