Simplify/restrict "advanced" 'creusot setup' workflows
Last commit makes 'creusot setup install' more flexible (with per-tool
flags). This commit removes some leftover bits of flexibility for
loading custom creusot and why3 configs, that may not be worth the
complexity.

This commit makes it so the creusot config is always looked up from
the "global" config file (removing the codepath specific to the
creusot git repository), and the why3 config file we use (why3.conf)
is always the one generated by creusot setup.