Fix purity validation
Don't confuse closures that are part of quantifiers with program closures

When `validate_purity` received a closure which is part of `forall` or a logic Mapping,
it was previously assigned purity `Program`, which is incorrect but that
mistake was masked because those closures are marked `no_translate`, which
causes `validate_purity` to skip the check.
However it can happen that `Purity::of_def_id` crashes on such closures;
I tried to isolate a test case when it happened to me and the crash went away,
that's why there is no test case...

The fix is to just skip `validate_purity` on all closures. Instead the check happens
while validating the purity of the parent function.

This fix also ensures that purity validation only traverses each expression once.