Mark Pearlite closures to avoid confusing them with regular closures
This avoids a crash in `validate_terminates`, specifically in `CallGraph::build`:
the `is_pearlite` check was wrongly returning `false` for closures of Pearlite quantifiers and mappings.
This was still fine most of the time because those closures are already marked `no_translate`, so they
are not traversed later anyway.
However it happened that the call to `sig` (right after `is_pearlite` returns `false`) crashes on such closures.