Weaken the opacity check.
For dependencies used in a proof_assert! statement in a logic definition,
or uses of FnDef values in logic definitions, we do not emit an error if
these dependencies are less visible than the transparency of the definition.
Typically, these dependencies are used to prove the ensures clauses of the
definition, and are not needed by users of the definition.

It is sound and composable, because these dependencies are ignored when
generating the Why3 code for the client.