Refactor the RA definitions.
The definitions are equivalent to the old ones, but (probably) more adapted to automated solvers.

My goal was to simplify the proof of fmap_view, which is made quite simpler.

The most important change is the refactoring of `maximal_idemp`.  The logic function `idemp` is removed, and the maximality of the core is now a separate lemma, which is not automatically loaded.