Make it possible to use `Deref` and `DerefMut` generically.
Some care has to be taken for Snapshot::deref(_mut) and
Ghost::deref(_mut), which should not be allowed to be called
generically. We add this restriction by adding a false precondition
either directly (for Ghost) or indirectly (for Snapshot, for which
deref and deref_mut are #[logic], and therefore there is an implicit
false precondition when used in generic contexts). This false
precondition is removed automatically when these methods are called
directly (and validation makes sure that no direct calls are possible
in invalid contexts).

In addition:
- Check opacity requirements in specification of functions (and fix
  all the code tat breaks because of this).
- Use #[creusot::no_translate] in a more rationnale way.
- A few simplification refactorings/bug fixes here and there.