Support for fully opaque types.
Generally, this simplify goals because there are fewer type definitions (i.e., we don't traverse the definition of Vec), but it may also lead to harder goals because some types had a trivial type invariant, which is no longer trivial because of opacity. Hence, this sometimes adds some VCs, which do not seem to be difficult in practice.